meta-intel: lms: Ensure do_unpack2 task can rerun
If something triggers a rebuild of the do_unpack2 task, it will fail if its
already run once. This patch ensures that it can be called multiple times.

(From meta-intel rev: b96f8e4c12f65b21851d00ea7de2bc53d03eafb2)

+++ meta-intel/common/recipes-bsp/amt/lms_7.1.20.bb
@@ -25,14 +25,16 @@ INITSCRIPT_PARAMS = "defaults"
 PV_SUB = "25"
 
 do_unpack2() {
-	     # The downloaded 37962 filename is actually lms+7.1.20.25.zip.
-	     # It contains lms-7.1.20-25.tar.gz.
-	     # It contains lms-7.1.20-25.tar.gz untars to lms-7.1.20
-	     mv ${WORKDIR}/37962 ${WORKDIR}/${PN}+${PV}.${PV_SUB}.zip
-	     unzip -o ${WORKDIR}/${PN}+${PV}.${PV_SUB}.zip
-	     mv ${WORKDIR}/${PN}-${PV}/outputdir/${PN}-${PV}-${PV_SUB}.tar.gz ${WORKDIR}/
-	     cd ${WORKDIR}
-	     tar -xvzf ${PN}-${PV}-${PV_SUB}.tar.gz
+	# The downloaded 37962 filename is actually lms+7.1.20.25.zip.
+	# It contains lms-7.1.20-25.tar.gz.
+	# It contains lms-7.1.20-25.tar.gz untars to lms-7.1.20
+	if [ -e "${WORKDIR}/37962" ]; then
+		mv ${WORKDIR}/37962 ${WORKDIR}/${PN}+${PV}.${PV_SUB}.zip
+		unzip -o ${WORKDIR}/${PN}+${PV}.${PV_SUB}.zip
+		mv ${WORKDIR}/${PN}-${PV}/outputdir/${PN}-${PV}-${PV_SUB}.tar.gz ${WORKDIR}/
+		cd ${WORKDIR}
+		tar -xvzf ${PN}-${PV}-${PV_SUB}.tar.gz
+	fi
 }
 
 addtask unpack2 after do_unpack before do_patch
